(#651, #573) - Remove old scenes before exporting.
Note that I've noticed another bug, which is that adding and removing the Profiler scene from BDX like we're doing is problematic if you delete the Profiler scene and re-export. Eventually BDX will error out, as Blender automatically re-names all of the objects to avoid using the same names as the previous Profiler scene (for some reason). The proper way to do this might be to use the operator to export the profiler scene from the profiler blend file (not to add the scene in to the current blend file before we export scenes, and then remove it afterwards).