(#654) - Fixing memory leak caused by filter downsampling.
Problem was caused by creation of new RenderBuffer for the tempBuffer whenever multiple ScreenShaders had different downsample values.

The solution is to remove the tempBuffer and create a map of screen sizes to render buffers. If a buffer for the screen size doesn't exist in the map, one is created and added to it. The tempBuffer RenderBuffer is replaced by getting the appropriate buffer from this map. When the screen size changes, the map is cleared (as the buffers may no longer have the same aspect ratio).