GameObject AABB enhancements.
aabbContains now has additional signatures that take individual numbers or a Vector3f.
aabbContains (for a single point) now can take an aabb for reference. If it is set to null, it will use the calling GameObject's AABB points.
aabbContains(float[][]) and aabbContainsAny(float[][]) now can take another aabb for reference. If it is set to null, it will use the calling GameObject's AABB points.
getAABBPoints() now can take a float value for an additional margin around the AABB box.